refactor: create dev app
