+++ packages/astro/test/units/vite-plugin-astro-server/request.test.js
@@ -4,8 +4,7 @@ import { createContainer } from '../../../dist/core/dev/container.js';
 import { createLoader } from '../../../dist/core/module-loader/index.js';
 import { createRoutesList } from '../../../dist/core/routing/index.js';
 import { createComponent, render } from '../../../dist/runtime/server/index.js';
-import { createController, handleRequest } from '../../../dist/vite-plugin-astro-server/index.js';
-import { DevPipeline } from '../../../dist/vite-plugin-astro-server/pipeline.js';
+import { createController, DevApp } from '../../../dist/vite-plugin-astro-server/index.js';
 import { createDevelopmentManifest } from '../../../dist/vite-plugin-astro-server/plugin.js';
 import testAdapter from '../../test-adapter.js';
 import {
@@ -16,7 +15,7 @@ import {
 	defaultLogger,
 } from '../test-utils.js';
 
-async function createDevPipeline(overrides = {}, root) {
+async function createDevApp(overrides = {}, root) {
 	const settings = overrides.settings ?? (await createBasicSettings({ root }));
 	const loader = overrides.loader ?? createLoader();
 	const manifest = createDevelopmentManifest(settings);
@@ -27,7 +26,8 @@ async function createDevPipeline(overrides = {}, root) {
 		},
 		defaultLogger,
 	);
-	return DevPipeline.create(routesList, { loader, logger: defaultLogger, manifest, settings });
+
+	return new DevApp(manifest, true, settings, defaultLogger, loader, routesList);
 }
 
 describe('vite-plugin-astro-server', () => {
@@ -38,7 +38,7 @@ describe('vite-plugin-astro-server', () => {
 				'/src/pages/index.astro': '',
 			});
 
-			const pipeline = await createDevPipeline(
+			const app = await createDevApp(
 				{
 					loader: createLoader({
 						import(id) {
@@ -54,17 +54,15 @@ describe('vite-plugin-astro-server', () => {
 				},
 				fixture.path,
 			);
-			const controller = createController({ loader: pipeline.loader });
+
+			const controller = createController({ loader: app.pipeline.loader });
 			const { req, res, text } = createRequestAndResponse();
 
 			try {
-				await handleRequest({
-					pipeline,
-					routesList: pipeline.routesList,
+				await app.handleRequest({
 					controller,
 					incomingRequest: req,
 					incomingResponse: res,
-					manifest: {},
 				});
 			} catch (err) {
 				assert.equal(err.message, undefined);
